requirements.txt
streamlit
python-dotenv
huggingface-hub
transformers
opencv-python
numpy
pandas
Pillow
hf_xet
torch
llama-cpp-python
google-generativeai
requests